[REFACTOR] : Hexagram 도메인 패키지 분리       - oracle.model 내부에 있던 Line, Lines, Hexagram, Trigram, HexagramMapping 이동       - domain/hexagram/model 패키지로 통합       - casting 패키지는 제거하고 hexagram 단일 도메인으로 단순화
