+++ src/main/java/com/hexaoracle/server/domain/oracle/service/ChangedHexagramService.java
@@ -1,6 +1,6 @@
 package com.hexaoracle.server.domain.oracle.service;
 
-import com.hexaoracle.server.domain.oracle.model.Lines;
+import com.hexaoracle.server.domain.hexagram.model.Lines;
 
 import java.util.ArrayList;
 import java.util.List;
